+++ src/main/java/wooteco/subway/exception/member/NotRegisteredMemberException.java
@@ -1,6 +1,7 @@
 package wooteco.subway.exception.member;
 
-public class NotRegisteredMemberException extends RuntimeException{
+public class NotRegisteredMemberException extends RuntimeException {
+
     public NotRegisteredMemberException() {
         super("[ERROR] 등록되지 않은 회원입니다.");
     }
