+++ src/main/java/wooteco/subway/exception/path/PathException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception.path;
+
+public class PathException extends RuntimeException {
+
+    public PathException(String message) {
+        super(message);
+    }
+}
