--- src/main/java/wooteco/subway/path/exceptions/PathException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.path.exceptions;
-
-public class PathException extends RuntimeException{
-
-    public PathException(String message) {
-        super(message);
-    }
-}
